Move designable into compoundForm, better naming
